Warnings: Make implicit conversions explicit casts

+++ src/fmu/Fmu1Value.c
@@ -265,7 +265,7 @@ Fmu1Value * Fmu1ValueReadArray(const char * logPrefix, ChannelType * type, Chann
     }
 
     for (i = startIdx; i <= endIdx; i++) {
-        char * indexedChannelName = CreateIndexedName(channelName, i);
+        char * indexedChannelName = CreateIndexedName(channelName, (unsigned int) i);
         fmi1_import_variable_t * var = fmi1_import_get_variable_by_name(fmiImport, indexedChannelName);
         if (!var) {
             mcx_log(LOG_ERROR, "%s: Could not get variable %s", logPrefix, indexedChannelName);
